+++ crates/project-graph/tests/project_graph_test.rs
@@ -1429,20 +1429,20 @@ mod project_graph {
         }
 
         #[tokio::test]
-        async fn by_task_platform() {
+        async fn by_task_toolchain() {
             let graph = generate_workspace_graph("query").await;
 
             let projects = graph
-                .query_projects(build_query("taskPlatform=[node]").unwrap())
+                .query_projects(build_query("taskToolchain=[node]").unwrap())
                 .unwrap();
 
-            assert_eq!(get_ids_from_projects(projects), vec!["a", "b"]);
+            assert_eq!(get_ids_from_projects(projects), vec!["a"]);
 
             let projects = graph
-                .query_projects(build_query("taskPlatform=system").unwrap())
+                .query_projects(build_query("taskToolchain=system").unwrap())
                 .unwrap();
 
-            assert_eq!(get_ids_from_projects(projects), vec!["c", "d"]);
+            assert_eq!(get_ids_from_projects(projects), vec!["a", "b", "c", "d"]);
         }
 
         #[tokio::test]
@@ -1461,7 +1461,7 @@ mod project_graph {
             let graph = generate_workspace_graph("query").await;
 
             let projects = graph
-                .query_projects(build_query("task=build && taskPlatform=deno").unwrap())
+                .query_projects(build_query("task=build && taskToolchain=deno").unwrap())
                 .unwrap();
 
             assert_eq!(get_ids_from_projects(projects), vec!["d"]);