+++ crates/workspace-graph/src/query_tasks.rs
@@ -85,8 +85,14 @@ fn does_task_match_criteria(task: &Task, query: &Criteria) -> miette::Result<boo
                         }
                     }
                     Field::Task(ids) => condition.matches(ids, &task.id),
-                    Field::TaskPlatform(platforms) => {
-                        condition.matches_enum(platforms, &task.platform)
+                    Field::TaskPlatform(ids) | Field::TaskToolchain(ids) => {
+                        let toolchains = task
+                            .toolchains
+                            .iter()
+                            .map(|t| t.as_str())
+                            .collect::<Vec<_>>();
+
+                        condition.matches_list(ids, &toolchains)
                     }
                     Field::TaskType(types) => condition.matches_enum(types, &task.type_of),
                     _ => Ok(false),